requirements.txt
google-genai>=1.11.0
rich>=13.7.0
PyYAML>=6.0.1
opencv-python>=4.9.0
numpy>=1.26.0
pyserial>=3.5
pillow>=10.0.0
pytest>=8.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-multipart>=0.0.9
